# Display help text when no arguments passed